feat(sampleApps): Update sample apps to React Native 0.80, update other dependencies, improve code quality (#5057)
* feat(sampleApps): update sample apps to React Native 0.80 / update other deps

* Lockfile update

* Make tests work
