+++ samples/react-native/.bundle/config
