+++ samples/react-native/ios/.xcode.env
@@ -7,7 +7,7 @@
 #
 # Customize the NODE_BINARY variable here.
 # For example, to use nvm with brew, add the following line
-#. "$(brew --prefix nvm)/nvm.sh" --no-use
+# . "$(brew --prefix nvm)/nvm.sh" --no-use
 export NODE_BINARY=$(command -v node)
 
 export EXTRA_COMPILER_ARGS="-w"
