+++ samples/react-native/metro.config.js
@@ -7,7 +7,7 @@ const { withMonorepo } = require('sentry-react-native-samples-utils/metro');
  * Metro configuration
  * https://reactnative.dev/docs/metro
  *
- * @type {import('metro-config').MetroConfig}
+ * @type {import('@react-native/metro-config').MetroConfig}
  */
 const config = {};
