+++ samples/react-native/package.json
@@ -22,45 +22,48 @@
     "build-android-release-splits-flavors": "export SENTRY_SAMPLE_ENABLE_ABI_SPLIT=true; export SENTRY_SAMPLE_ENABLE_FLAVORS=true; cd android; ./gradlew assembleRelease; cd .."
   },
   "dependencies": {
-    "@react-navigation/bottom-tabs": "7.3.12",
-    "@react-navigation/native": "7.1.8",
-    "@react-navigation/native-stack": "7.3.12",
-    "@react-navigation/stack": "7.3.1",
+    "@react-native-vector-icons/ionicons": "^12.3.0",
+    "@react-native/new-app-screen": "0.80.2",
+    "@react-navigation/bottom-tabs": "^7.4.5",
+    "@react-navigation/native": "^7.1.17",
+    "@react-navigation/native-stack": "^7.3.24",
+    "@react-navigation/stack": "^7.4.5",
+    "@reduxjs/toolkit": "^2.8.2",
     "@sentry/core": "10.12.0",
+    "@sentry/react": "^10.2.0",
     "@sentry/react-native": "7.1.0",
-    "@shopify/flash-list": "1.8.0",
+    "@shopify/flash-list": "^2.0.2",
     "delay": "^6.0.0",
-    "react": "19.0.0",
-    "react-native": "0.79.2",
-    "react-native-gesture-handler": "^2.24.0",
-    "react-native-image-picker": "^8.0.0",
+    "react": "19.1.0",
+    "react-native": "0.80.2",
+    "react-native-gesture-handler": "^2.28.0",
+    "react-native-image-picker": "^8.2.1",
     "react-native-launch-arguments": "^4.1.0",
-    "react-native-reanimated": "3.17.1",
-    "react-native-safe-area-context": "5.2.0",
-    "react-native-screens": "4.10.0",
-    "react-native-svg": "15.11.2",
-    "react-native-vector-icons": "^10.2.0",
-    "react-redux": "^8.1.3",
-    "redux": "^4.2.1",
+    "react-native-reanimated": "3.19.1",
+    "react-native-safe-area-context": "^5.5.2",
+    "react-native-screens": "^4.13.1",
+    "react-native-svg": "^15.12.1",
+    "react-native-webview": "^13.15.0",
+    "react-redux": "^9.2.0",
     "setimmediate": "^1.0.5"
   },
   "devDependencies": {
     "@babel/core": "^7.26.7",
     "@babel/preset-env": "^7.26.7",
     "@babel/runtime": "^7.26.7",
-    "@react-native-community/cli": "15.1.3",
-    "@react-native-community/cli-platform-android": "15.1.3",
-    "@react-native-community/cli-platform-ios": "15.1.3",
-    "@react-native/babel-preset": "0.77.1",
-    "@react-native/eslint-config": "0.77.1",
-    "@react-native/metro-config": "0.77.1",
-    "@react-native/typescript-config": "0.77.1",
+    "@react-native-community/cli": "19.1.1",
+    "@react-native-community/cli-platform-android": "19.1.1",
+    "@react-native-community/cli-platform-ios": "19.1.1",
+    "@react-native/babel-preset": "0.80.2",
+    "@react-native/eslint-config": "0.80.2",
+    "@react-native/metro-config": "0.80.2",
+    "@react-native/typescript-config": "0.80.2",
     "@sentry/babel-plugin-component-annotate": "4.3.0",
     "@testing-library/react-native": "^13.2.2",
     "@types/jest": "^29.5.14",
     "@types/node": "^22.13.1",
-    "@types/react": "^19.0.0",
-    "@types/react-native-vector-icons": "^6.4.18",
+    "@types/react": "^19.1.0",
+    "@types/react-test-renderer": "^19.1.0",
     "@typescript-eslint/eslint-plugin": "^7.18.0",
     "@typescript-eslint/parser": "^7.18.0",
     "babel-jest": "^29.6.3",
@@ -70,13 +73,17 @@
     "jest": "^29.6.3",
     "patch-package": "^8.0.0",
     "prettier": "2.8.8",
-    "react-test-renderer": "19.0.0",
+    "react-test-renderer": "19.1.0",
     "sentry-react-native-samples-utils": "workspace:^",
     "ts-jest": "^29.3.1",
     "typescript": "5.0.4"
   },
   "engines": {
     "node": ">=18"
+  },
+  "volta": {
+    "node": "18.20.8",
+    "yarn": "3.6.4"
   },
   "codegenConfig": {
     "name": "AppSpecs",
