+++ samples/react-native/src/App.tsx
@@ -1,54 +1,32 @@
 import React from 'react';
+
+import { Ionicons } from '@react-native-vector-icons/ionicons';
+import { createBottomTabNavigator } from '@react-navigation/bottom-tabs';
 import {
   NavigationContainer,
   NavigationContainerRef,
+  TypedNavigator,
 } from '@react-navigation/native';
 import { createNativeStackNavigator } from '@react-navigation/native-stack';
 import { createStackNavigator } from '@react-navigation/stack';
-import { createBottomTabNavigator } from '@react-navigation/bottom-tabs';
-import Animated, {
-  Easing,
-  useAnimatedStyle,
-  useSharedValue,
-  withRepeat,
-  withTiming,
-} from 'react-native-reanimated';
-
-// Import the Sentry React Native SDK
 import * as Sentry from '@sentry/react-native';
-// eslint-disable-next-line @typescript-eslint/no-unused-vars
-import { withSentryPlayground } from '@sentry/react-native/playground';
-import { FeedbackWidget } from '@sentry/react-native';
-
-import ErrorsScreen from './Screens/ErrorsScreen';
-import PerformanceScreen from './Screens/PerformanceScreen';
-import TrackerScreen from './Screens/TrackerScreen';
-import ManualTrackerScreen from './Screens/ManualTrackerScreen';
-import PerformanceTimingScreen from './Screens/PerformanceTimingScreen';
-import ReduxScreen from './Screens/ReduxScreen';
-import { Provider } from 'react-redux';
-import { store } from './reduxApp';
-import { GestureHandlerRootView } from 'react-native-gesture-handler';
-import GesturesTracingScreen from './Screens/GesturesTracingScreen';
-import { LogBox, Platform, StyleSheet, View } from 'react-native';
-import Ionicons from 'react-native-vector-icons/Ionicons';
-import PlaygroundScreen from './Screens/PlaygroundScreen';
-import { getDsn, logWithoutTracing } from './utils';
-import HeavyNavigationScreen from './Screens/HeavyNavigationScreen';
-import WebviewScreen from './Screens/WebviewScreen';
 import { isTurboModuleEnabled } from '@sentry/react-native/dist/js/utils/environment';
+import { LogBox, Platform } from 'react-native';
 import * as ImagePicker from 'react-native-image-picker';
-import SpaceflightNewsScreen from './Screens/SpaceflightNewsScreen';
 
-/* false by default to avoid issues in e2e tests waiting for the animation end */
-const RUNNING_INDICATOR = false;
+import RunningIndicator from './components/RunningIndicator';
+import WebviewScreen from './Screens/WebviewScreen';
+import getErrorsTab from './tabs/ErrorsTab';
+import getPerformanceTab from './tabs/PerformanceTab';
+import getPlaygroundTab from './tabs/PlaygroundTab';
+import { getDsn, logWithoutTracing } from './utils';
+
+LogBox.ignoreAllLogs();
+const isMobileOs = Platform.OS === 'android' || Platform.OS === 'ios';
 
 if (typeof setImmediate === 'undefined') {
   require('setimmediate');
 }
-
-LogBox.ignoreAllLogs();
-const isMobileOs = Platform.OS === 'android' || Platform.OS === 'ios';
 
 const reactNavigationIntegration = Sentry.reactNavigationIntegration({
   routeChangeTimeoutMs: 500, // How long it will wait for the route change to complete. Default is 1000ms
@@ -60,6 +38,11 @@ const reactNavigationIntegration = Sentry.reactNavigationIntegration({
 
 const sampleFeatureFlagsIntegration = Sentry.featureFlagsIntegration();
 sampleFeatureFlagsIntegration.addFeatureFlag('sample-test-flag', true);
+
+const StackNavigator: TypedNavigator<any, any> = isMobileOs
+  ? createNativeStackNavigator()
+  : createStackNavigator();
+const BottomTabNavigator = createBottomTabNavigator();
 
 Sentry.init({
   // Replace the example DSN below with your own DSN:
@@ -120,7 +103,7 @@ Sentry.init({
           Platform.OS === 'ios' && isTurboModuleEnabled()
             ? // The global patch doesn't work on iOS with the New Architecture in this Sample app
               // In
-              false
+            false
             : true,
       }),
       Sentry.feedbackIntegration({
@@ -184,101 +167,16 @@ Sentry.init({
   autoInitializeNativeSdk: true,
 });
 
-const Stack = isMobileOs
-  ? createNativeStackNavigator()
-  : createStackNavigator();
-const Tab = createBottomTabNavigator();
-
-const ErrorsTabNavigator = Sentry.withProfiler(
-  () => {
-    return (
-      <GestureHandlerRootView style={styles.wrapper}>
-        <Provider store={store}>
-          <Stack.Navigator>
-            <Stack.Screen
-              name="ErrorsScreen"
-              component={ErrorsScreen}
-              options={{ title: 'Errors' }}
-            />
-            <Stack.Screen
-              name="FeedbackWidget"
-              options={{ presentation: 'modal', headerShown: false }}>
-              {props => (
-                <FeedbackWidget
-                  {...props}
-                  enableScreenshot={true}
-                  onFormClose={props.navigation.goBack}
-                  onFormSubmitted={props.navigation.goBack}
-                  styles={{
-                    submitButton: {
-                      backgroundColor: '#6a1b9a',
-                      paddingVertical: 15,
-                      borderRadius: 5,
-                      alignItems: 'center',
-                      marginBottom: 10,
-                    },
-                  }}
-                  namePlaceholder={'Fullname'}
-                />
-              )}
-            </Stack.Screen>
-          </Stack.Navigator>
-        </Provider>
-      </GestureHandlerRootView>
-    );
-  },
-  { name: 'ErrorsTab' },
-);
-
-const PerformanceTabNavigator = Sentry.withProfiler(
-  () => {
-    return (
-      <GestureHandlerRootView style={styles.wrapper}>
-        <Provider store={store}>
-          <Stack.Navigator>
-            <Stack.Screen
-              name="PerformanceScreen"
-              component={PerformanceScreen}
-              options={{ title: 'Performance' }}
-            />
-            <Stack.Screen
-              name="SpaceflightNewsScreen"
-              component={SpaceflightNewsScreen}
-            />
-            <Stack.Screen name="Tracker" component={TrackerScreen} />
-            <Stack.Screen
-              name="ManualTracker"
-              component={ManualTrackerScreen}
-            />
-            <Stack.Screen
-              name="HeavyNavigation"
-              component={HeavyNavigationScreen}
-            />
-            <Stack.Screen
-              name="PerformanceTiming"
-              component={PerformanceTimingScreen}
-            />
-            <Stack.Screen name="Redux" component={ReduxScreen} />
-            <Stack.Screen name="Gestures" component={GesturesTracingScreen} />
-          </Stack.Navigator>
-        </Provider>
-      </GestureHandlerRootView>
-    );
-  },
-  { name: 'PerformanceTab' },
-);
-
 function BottomTabsNavigator() {
   return (
-    <Tab.Navigator
+    <BottomTabNavigator.Navigator
       screenOptions={{
         headerShown: false,
       }}
-      detachInactiveScreens={false} // workaround for https://github.com/react-navigation/react-navigation/issues/11384
-    >
-      <Tab.Screen
+      detachInactiveScreens={false}>
+      <BottomTabNavigator.Screen
         name="ErrorsTab"
-        component={ErrorsTabNavigator}
+        component={getErrorsTab(StackNavigator)}
         options={{
           tabBarLabel: 'Errors',
           // eslint-disable-next-line react/no-unstable-nested-components
@@ -292,9 +190,9 @@ function BottomTabsNavigator() {
           ),
         }}
       />
-      <Tab.Screen
+      <BottomTabNavigator.Screen
         name="PerformanceTab"
-        component={PerformanceTabNavigator}
+        component={getPerformanceTab(StackNavigator)}
         options={{
           tabBarLabel: 'Performance',
           // eslint-disable-next-line react/no-unstable-nested-components
@@ -308,9 +206,9 @@ function BottomTabsNavigator() {
           ),
         }}
       />
-      <Tab.Screen
+      <BottomTabNavigator.Screen
         name="PlaygroundTab"
-        component={PlaygroundScreen}
+        component={getPlaygroundTab()}
         options={{
           tabBarLabel: 'Playground',
           // eslint-disable-next-line react/no-unstable-nested-components
@@ -324,7 +222,7 @@ function BottomTabsNavigator() {
           ),
         }}
       />
-    </Tab.Navigator>
+    </BottomTabNavigator.Navigator>
   );
 }
 
@@ -337,17 +235,20 @@ function RootNavigationContainer() {
       onReady={() => {
         reactNavigationIntegration.registerNavigationContainer(navigation);
       }}>
-      <Stack.Navigator
+      <StackNavigator.Navigator
         screenOptions={{
           headerShown: false,
         }}>
-        <Stack.Screen name="BottomTabs" component={BottomTabsNavigator} />
-        <Stack.Screen
+        <StackNavigator.Screen
+          name="BottomTabs"
+          component={BottomTabsNavigator}
+        />
+        <StackNavigator.Screen
           name="Webview"
           component={WebviewScreen}
           options={{ headerShown: true }}
         />
-      </Stack.Navigator>
+      </StackNavigator.Navigator>
     </NavigationContainer>
   );
 }
@@ -360,64 +261,5 @@ function App() {
     </>
   );
 }
-
-function RunningIndicator() {
-  if (Platform.OS !== 'android' && Platform.OS !== 'ios') {
-    return null;
-  }
-
-  if (!RUNNING_INDICATOR) {
-    return null;
-  }
-
-  return <RotatingBox />;
-}
-
-function RotatingBox() {
-  const sv = useSharedValue<number>(0);
-
-  React.useEffect(() => {
-    sv.value = withRepeat(
-      withTiming(360, {
-        duration: 1_000_000,
-        easing: Easing.linear,
-      }),
-      -1,
-    );
-    // eslint-disable-next-line react-hooks/exhaustive-deps
-  }, []);
-
-  const animatedStyle = useAnimatedStyle(() => ({
-    transform: [{ rotate: `${sv.value * 360}deg` }],
-  }));
-
-  return (
-    <View style={styles.container}>
-      <Animated.View style={[styles.box, animatedStyle]} />
-    </View>
-  );
-}
-
-const styles = StyleSheet.create({
-  wrapper: {
-    flex: 1,
-  },
-  container: {
-    position: 'absolute',
-    left: 30,
-    top: 30,
-  },
-  box: {
-    height: 50,
-    width: 50,
-    backgroundColor: '#b58df1',
-    borderRadius: 5,
-  },
-});
-
-// export default withSentryPlayground(Sentry.wrap(App), {
-//   projectId: '5428561',
-//   organizationSlug: 'sentry-sdks',
-// });
 
 export default Sentry.wrap(App);
