+++ samples/react-native/src/Screens/ReduxScreen.tsx
@@ -1,5 +1,5 @@
 import * as React from 'react';
-import { View, StyleSheet } from 'react-native';
+import { StyleSheet, View } from 'react-native';
 
 import Counter from '../components/Counter';
