+++ samples/react-native/src/components/RunningIndicator.tsx
@@ -1,0 +1,63 @@
+import React from 'react';
+import { Platform, StyleSheet, View } from 'react-native';
+import Animated, {
+  Easing,
+  useAnimatedStyle,
+  useSharedValue,
+  withRepeat,
+  withTiming,
+} from 'react-native-reanimated';
+
+/* false by default to avoid issues in e2e tests waiting for the animation end */
+const RUNNING_INDICATOR = false;
+
+function RotatingBox() {
+  const sv = useSharedValue<number>(0);
+
+  React.useEffect(() => {
+    sv.value = withRepeat(
+      withTiming(360, {
+        duration: 1_000_000,
+        easing: Easing.linear,
+      }),
+      -1,
+    );
+    // eslint-disable-next-line react-hooks/exhaustive-deps
+  }, []);
+
+  const animatedStyle = useAnimatedStyle(() => ({
+    transform: [{ rotate: `${sv.value * 360}deg` }],
+  }));
+
+  return (
+    <View style={styles.container}>
+      <Animated.View style={[styles.box, animatedStyle]} />
+    </View>
+  );
+}
+
+export default function RunningIndicator() {
+  if (Platform.OS !== 'android' && Platform.OS !== 'ios') {
+    return null;
+  }
+
+  if (!RUNNING_INDICATOR) {
+    return null;
+  }
+
+  return <RotatingBox />;
+}
+
+const styles = StyleSheet.create({
+  container: {
+    position: 'absolute',
+    left: 30,
+    top: 30,
+  },
+  box: {
+    height: 50,
+    width: 50,
+    backgroundColor: '#b58df1',
+    borderRadius: 5,
+  },
+});
